Creating one file for each aggregation element.
