+++ graylog2-web-interface/src/views/components/aggregationwizard/aggregationElements/AggregationElementType.ts
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2020 Graylog, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Server Side Public License, version 1,
+ * as published by MongoDB, Inc.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Server Side Public License for more details.
+ *
+ * You should have received a copy of the Server Side Public License
+ * along with this program. If not, see
+ * <http://www.mongodb.com/licensing/server-side-public-license>.
+ */
+import AggregationWidgetConfig from 'views/logic/aggregationbuilder/AggregationWidgetConfig';
+
+import type { WidgetConfigFormValues } from '../WidgetConfigForm';
+
+export type AggregationElement = {
+  title: string,
+  key: string,
+  isConfigured: (formValues: WidgetConfigFormValues) => boolean,
+  multipleUse: boolean,
+  order: number,
+  toConfig?: (formValues: WidgetConfigFormValues, currentConfig: AggregationWidgetConfig) => AggregationWidgetConfig,
+  fromConfig?: (config: AggregationWidgetConfig) => Partial<WidgetConfigFormValues>,
+  onCreate?: () => void,
+  onDeleteAll?: (formValues: WidgetConfigFormValues) => WidgetConfigFormValues,
+  component: React.ComponentType<{
+    config: AggregationWidgetConfig,
+    onConfigChange: (newConfig: AggregationWidgetConfig) => void
+  }>,
+  validate?: (formValues: WidgetConfigFormValues) => { [key: string]: string },
+}
